+++ NEWS.md
@@ -1,7 +1,7 @@
 Version 0.1.1
 ================
 
-* Added flexibility in plotting appearance: axes labels and legend titles can now be customized.
+* Added flexibility in plotting appearance: axis labels and legend titles can now be customized.
 
 
 Version 0.1.0