+++ NEWS.md
@@ -2,6 +2,7 @@ Version 0.1.1
 ================
 
 * Added flexibility in plotting appearance: axis labels and legend titles can now be customized.
+* Includes deSolve error messages if equation solving fails
 
 
 Version 0.1.0